changes for proper comparison
