+++ pom.xml
@@ -12,7 +12,7 @@
         <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <matsim.version>16.0-PR3034</matsim.version>
+        <matsim.version>16.0-PR3096</matsim.version>
     </properties>
