+++ src/main/java/org/sinfonica/inclusive_ccam/heterogenous_users/drt/UserSpecificStopTimeModule.java
@@ -14,8 +14,8 @@ public class UserSpecificStopTimeModule extends AbstractDvrpModeModule {
     @Override
     public void install() {
         bindModal(PassengerStopDurationProvider.class).to(UserSpecificStopTimeProvider.class);
-        bindModal(ParallelStopTimeCalculator.class).toProvider(modalProvider(getter -> new ParallelStopTimeCalculator(getter.getModal(PassengerStopDurationProvider.class))));
-        bindModal(StopTimeCalculator.class).to(modalKey(ParallelStopTimeCalculator.class));
+        // bindModal(ParallelStopTimeCalculator.class).toProvider(modalProvider(getter -> new ParallelStopTimeCalculator(getter.getModal(PassengerStopDurationProvider.class))));
+        // bindModal(StopTimeCalculator.class).to(modalKey(ParallelStopTimeCalculator.class));
     }
 
 }
